Separate handlers and healthcheck tests
